Updated callbacks and added getUserMedia error handling
