+++ webrtc/videoroom/assets/css/app.scss
@@ -39,6 +39,7 @@ h3 {
     text-decoration: none;
     cursor: pointer;
 
+    &,
     &:visited,
     &:active {
       color: inherit;
