+++ webrtc/videoroom/assets/src/app.ts
@@ -1,60 +1,46 @@
 import "../css/app.scss";
-//
-import "phoenix_html";
+
+import { addVideoElement, removeVideoElement, setErrorMessage } from "./ui";
 
 import { MembraneWebRTC } from "./membraneWebRTC";
 import { Socket } from "phoenix";
 
-function addVideoElement(
-  stream: MediaStream,
-  _: MediaStreamTrack,
-  mute: boolean = false
-) {
-  let video = <HTMLVideoElement>document.getElementById(stream.id);
+const CONSTRAINTS = {
+  audio: true,
+  video: { width: 1280, height: 720 },
+};
 
-  if (!video) {
-    video = document.createElement("video");
-    video.id = stream.id;
-    document.getElementById("videochat")?.appendChild(video);
+const setup = async () => {
+  const socket = new Socket("/socket");
+  socket.connect();
+
+  const roomEl = document.getElementById("room");
+  if (roomEl) {
+    try {
+      const roomId = roomEl.dataset.roomId || "lobby";
+
+      const localStream = await navigator.mediaDevices.getUserMedia(
+        CONSTRAINTS
+      );
+
+      localStream
+        .getTracks()
+        .forEach((track) => addVideoElement(track, localStream, true));
+
+      new MembraneWebRTC(socket, localStream, roomId, {
+        onAddTrack: addVideoElement,
+        onRemoveTrack: removeVideoElement,
+        onConnectionError: setErrorMessage,
+      }).start();
+    } catch (error) {
+      console.error(error);
+      setErrorMessage(
+        "Failed to setup video room, make sure to grant camera and microphone permissions"
+      );
+    }
+  } else {
+    console.error("room element is missing, cannot join video room");
   }
-  video.srcObject = stream;
-  video.autoplay = true;
-  video.playsInline = true;
-  video.muted = mute;
-}
+};
 
-function removeVideoElement(stream: MediaStream, _: MediaStreamTrack) {
-  if (stream.getTracks().length > 0) {
-    return;
-  }
-
-  const video = <HTMLVideoElement>document.getElementById(stream.id);
-  if (video) {
-    video.remove();
-  }
-}
-
-function setErrorMessage(
-  message: string = "Cannot connect to server, refresh the page and try again"
-) {
-  const control = document.getElementById("control");
-  if (control) {
-    control.innerHTML = message;
-  }
-}
-
-const socket = new Socket("/socket");
-socket.connect();
-
-const roomEl = document.getElementById("room");
-if (roomEl) {
-  const roomId = roomEl.dataset.roomId || "lobby";
-
-  new MembraneWebRTC(socket, roomId, {
-    onAddTrack: addVideoElement,
-    onRemoveTrack: removeVideoElement,
-    onConnectionError: setErrorMessage,
-  }).start();
-} else {
-  console.error("room element is missing, cannot join video room");
-}
+setup();
